[WebView2] namepsace refactoring
