+++ src/client/Microsoft.Identity.Client/Platforms/Features/WinFormsLegacyWebUi/SilentWindowsFormsAuthenticationDialog.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Runtime.InteropServices;
 using System.Windows.Forms;
 
-namespace Microsoft.Identity.Client.Platforms.net45
+namespace Microsoft.Identity.Client.Platforms.Features.WinFormsLegacyWebUi
 {
     /// <summary>
     /// </summary>
